requirements.txt
streamlit==1.52.1
python-pptx==1.0.2
google-genai